Monad rewrite
Created a simple State monad and implemented a GameState monad on top of it with the game's Engine type. This allows much nicer syntax for progressing the game (instead of the previous method of taking in an Engine, modifying it, outputting it, and catching the new Engine to process progress).

Also added lots of documentation and linted all the code.